vm: move witness-related types from operations module
