+++ src/operation/seal.rs
@@ -26,23 +26,18 @@ use std::hash::Hash;
 use bp::dbc::Method;
 pub use bp::seals::txout::blind::{ChainBlindSeal, ParseError, SingleBlindSeal};
 pub use bp::seals::txout::TxoSeal;
-use bp::seals::txout::{BlindSeal, CloseMethod, ExplicitSeal, SealTxid, VerifyError, Witness};
+use bp::seals::txout::{BlindSeal, CloseMethod, ExplicitSeal, SealTxid};
 pub use bp::seals::SecretSeal;
-use bp::{dbc, Outpoint, Tx, Txid, Vout};
-use commit_verify::{mpc, Conceal};
-use single_use_seals::SealWitness;
+use bp::{Outpoint, Txid, Vout};
+use commit_verify::Conceal;
 use strict_encoding::{StrictDecode, StrictDumb, StrictEncode};
 
-use crate::operation::xchain::Impossible;
 use crate::{XChain, XOutpoint};
 
 pub type GenesisSeal = SingleBlindSeal<Method>;
 pub type GraphSeal = ChainBlindSeal<Method>;
 
 pub type OutputSeal = ExplicitSeal<Txid, Method>;
-
-// TODO: Move to vm::Contract
-pub type XWitnessId = XChain<Txid>;
 
 pub type XGenesisSeal = XChain<GenesisSeal>;
 pub type XGraphSeal = XChain<GraphSeal>;
@@ -133,97 +128,6 @@ impl XChain<GenesisSeal> {
     /// Converts seal into a transaction outpoint.
     #[inline]
     pub fn to_outpoint(&self) -> XOutpoint { self.map_ref(GenesisSeal::to_outpoint).into() }
-}
-
-impl<U: ExposedSeal> XChain<U> {
-    pub fn method(self) -> CloseMethod
-    where U: TxoSeal {
-        match self {
-            XChain::Bitcoin(seal) => seal.method(),
-            XChain::Liquid(seal) => seal.method(),
-            XChain::Other(_) => unreachable!(),
-        }
-    }
-
-    #[inline]
-    pub fn to_output_seal(self) -> Option<XOutputSeal>
-    where U: TxoSeal {
-        Some(match self {
-            XChain::Bitcoin(seal) => {
-                let outpoint = seal.outpoint()?;
-                XChain::Bitcoin(ExplicitSeal::new(seal.method(), outpoint))
-            }
-            XChain::Liquid(seal) => {
-                let outpoint = seal.outpoint()?;
-                XChain::Liquid(ExplicitSeal::new(seal.method(), outpoint))
-            }
-            XChain::Other(_) => unreachable!(),
-        })
-    }
-
-    pub fn try_to_output_seal(self, witness_id: XWitnessId) -> Result<XOutputSeal, Self>
-    where U: TxoSeal {
-        self.to_output_seal()
-            .or(match (self, witness_id) {
-                (XChain::Bitcoin(seal), XWitnessId::Bitcoin(txid)) => {
-                    Some(XChain::Bitcoin(ExplicitSeal::new(seal.method(), seal.outpoint_or(txid))))
-                }
-                (XChain::Liquid(seal), XWitnessId::Liquid(txid)) => {
-                    Some(XChain::Liquid(ExplicitSeal::new(seal.method(), seal.outpoint_or(txid))))
-                }
-                _ => None,
-            })
-            .ok_or(self)
-    }
-}
-
-// TODO: Move to vm::Contract
-pub type XWitnessTx<X = Impossible> = XChain<Tx, X>;
-
-impl XWitnessTx {
-    pub fn witness_id(&self) -> XWitnessId {
-        match self {
-            Self::Bitcoin(tx) => XWitnessId::Bitcoin(tx.txid()),
-            Self::Liquid(tx) => XWitnessId::Liquid(tx.txid()),
-            Self::Other(_) => unreachable!(),
-        }
-    }
-}
-
-impl<Dbc: dbc::Proof> XChain<Witness<Dbc>> {
-    pub fn witness_id(&self) -> XWitnessId {
-        match self {
-            Self::Bitcoin(w) => XWitnessId::Bitcoin(w.txid),
-            Self::Liquid(w) => XWitnessId::Liquid(w.txid),
-            Self::Other(_) => unreachable!(),
-        }
-    }
-}
-
-impl<Dbc: dbc::Proof, Seal: TxoSeal> SealWitness<Seal> for XChain<Witness<Dbc>> {
-    type Message = mpc::Commitment;
-    type Error = VerifyError<Dbc::Error>;
-
-    fn verify_seal(&self, seal: &Seal, msg: &Self::Message) -> Result<(), Self::Error> {
-        match self {
-            Self::Bitcoin(witness) | Self::Liquid(witness) => witness.verify_seal(seal, msg),
-            Self::Other(_) => unreachable!(),
-        }
-    }
-
-    fn verify_many_seals<'seal>(
-        &self,
-        seals: impl IntoIterator<Item = &'seal Seal>,
-        msg: &Self::Message,
-    ) -> Result<(), Self::Error>
-    where
-        Seal: 'seal,
-    {
-        match self {
-            Self::Bitcoin(witness) | Self::Liquid(witness) => witness.verify_many_seals(seals, msg),
-            Self::Other(_) => unreachable!(),
-        }
-    }
 }
 
 impl<Id: SealTxid> XChain<BlindSeal<Id>> {
