+++ src/vm/contract.rs
@@ -29,14 +29,111 @@ use std::rc::Rc;
 
 use amplify::num::u24;
 use amplify::Bytes32;
+use bp::seals::txout::{CloseMethod, ExplicitSeal, VerifyError, Witness};
+use bp::{dbc, Tx, Txid};
+use commit_verify::mpc;
+use single_use_seals::SealWitness;
 use strict_encoding::{StrictDecode, StrictDumb, StrictEncode};
 
 use crate::{
     AssetTags, AssignmentType, Assignments, AssignmentsRef, AttachState, ContractId, DataState,
-    Extension, ExtensionType, FungibleState, Genesis, GlobalState, GlobalStateType, GraphSeal,
-    Inputs, Metadata, OpFullType, OpId, OpType, Operation, Transition, TransitionType,
-    TypedAssigns, Valencies, XOutpoint, XWitnessId, LIB_NAME_RGB_LOGIC,
+    ExposedSeal, Extension, ExtensionType, FungibleState, Genesis, GlobalState, GlobalStateType,
+    GraphSeal, Impossible, Inputs, Metadata, OpFullType, OpId, OpType, Operation, Transition,
+    TransitionType, TxoSeal, TypedAssigns, Valencies, XChain, XOutpoint, XOutputSeal,
+    LIB_NAME_RGB_LOGIC,
 };
+
+pub type XWitnessId = XChain<Txid>;
+
+pub type XWitnessTx<X = Impossible> = XChain<Tx, X>;
+
+impl XWitnessTx {
+    pub fn witness_id(&self) -> XWitnessId {
+        match self {
+            Self::Bitcoin(tx) => XWitnessId::Bitcoin(tx.txid()),
+            Self::Liquid(tx) => XWitnessId::Liquid(tx.txid()),
+            Self::Other(_) => unreachable!(),
+        }
+    }
+}
+
+impl<Dbc: dbc::Proof> XChain<Witness<Dbc>> {
+    pub fn witness_id(&self) -> XWitnessId {
+        match self {
+            Self::Bitcoin(w) => XWitnessId::Bitcoin(w.txid),
+            Self::Liquid(w) => XWitnessId::Liquid(w.txid),
+            Self::Other(_) => unreachable!(),
+        }
+    }
+}
+
+impl<Dbc: dbc::Proof, Seal: TxoSeal> SealWitness<Seal> for XChain<Witness<Dbc>> {
+    type Message = mpc::Commitment;
+    type Error = VerifyError<Dbc::Error>;
+
+    fn verify_seal(&self, seal: &Seal, msg: &Self::Message) -> Result<(), Self::Error> {
+        match self {
+            Self::Bitcoin(witness) | Self::Liquid(witness) => witness.verify_seal(seal, msg),
+            Self::Other(_) => unreachable!(),
+        }
+    }
+
+    fn verify_many_seals<'seal>(
+        &self,
+        seals: impl IntoIterator<Item = &'seal Seal>,
+        msg: &Self::Message,
+    ) -> Result<(), Self::Error>
+    where
+        Seal: 'seal,
+    {
+        match self {
+            Self::Bitcoin(witness) | Self::Liquid(witness) => witness.verify_many_seals(seals, msg),
+            Self::Other(_) => unreachable!(),
+        }
+    }
+}
+
+impl<U: ExposedSeal> XChain<U> {
+    pub fn method(self) -> CloseMethod
+    where U: TxoSeal {
+        match self {
+            XChain::Bitcoin(seal) => seal.method(),
+            XChain::Liquid(seal) => seal.method(),
+            XChain::Other(_) => unreachable!(),
+        }
+    }
+
+    #[inline]
+    pub fn to_output_seal(self) -> Option<XOutputSeal>
+    where U: TxoSeal {
+        Some(match self {
+            XChain::Bitcoin(seal) => {
+                let outpoint = seal.outpoint()?;
+                XChain::Bitcoin(ExplicitSeal::new(seal.method(), outpoint))
+            }
+            XChain::Liquid(seal) => {
+                let outpoint = seal.outpoint()?;
+                XChain::Liquid(ExplicitSeal::new(seal.method(), outpoint))
+            }
+            XChain::Other(_) => unreachable!(),
+        })
+    }
+
+    pub fn try_to_output_seal(self, witness_id: XWitnessId) -> Result<XOutputSeal, Self>
+    where U: TxoSeal {
+        self.to_output_seal()
+            .or(match (self, witness_id) {
+                (XChain::Bitcoin(seal), XWitnessId::Bitcoin(txid)) => {
+                    Some(XChain::Bitcoin(ExplicitSeal::new(seal.method(), seal.outpoint_or(txid))))
+                }
+                (XChain::Liquid(seal), XWitnessId::Liquid(txid)) => {
+                    Some(XChain::Liquid(ExplicitSeal::new(seal.method(), seal.outpoint_or(txid))))
+                }
+                _ => None,
+            })
+            .ok_or(self)
+    }
+}
 
 #[derive(Copy, Clone, PartialEq, Eq, Debug, From)]
 pub enum OpRef<'op> {
